@@ -21,11 +21,11 @@
     "antd": "^4.19.5",
     "axios": "^0.27.2",
     "firebase": "^8.8.0",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.1.0",
+    "react-dom": "^18.1.0",
     "react-icons": "^4.2.0",
     "react-loading": "^2.0.3",
-    "react-router-dom": "^5.2.0",
+    "react-router-dom": "6",
     "styled-components": "^5.3.0",
     "vite": "^2.9.5",
     "web-vitals": "^1.0.1"
