@@ -1,15 +1,15 @@
-import React, { useEffect, useState, useContext } from 'react';
+import React, { useEffect, useState } from 'react';
 import logo from '../../assets/images/logo.svg';
 import './LoginScreen.css';
-import { Redirect } from 'react-router-dom';
+import { useNavigate } from 'react-router-dom';
 import ReactLoading from 'react-loading';
-import { UserContext } from '../../contexts/userContext';
+import { useUser } from '../../contexts/userContext';
 
 import { useAuth } from '../../firebase'
 
 const LoginScreen = () => {
   const [email, setEmail] = useState('');
-  const { setUserInfo } = useContext(UserContext);
+  const { userInfo, setUserInfo } = useUser();
   const [password, setPassword] = useState('');
   const [emailError, setEmailError] = useState('');
   const [passwordError, setPasswordError] = useState('');
@@ -24,10 +24,12 @@ const LoginScreen = () => {
     setPasswordError('');
   }
 
-  const handleLogin = () => {
+  const handleLogin = (e) => {
+    e.preventDefault()
     setIsLoading(true);
     clearErrors();
     signIn(email, password).then((data) => {
+      console.log(data)
       switch (data) {
         case "auth/invalid-email":
         case "auth/user-disabled":
@@ -40,6 +42,7 @@ const LoginScreen = () => {
           setIsLoading(false);
           break;
         default:
+          setUserInfo(data)
           setShouldRedirectToApp(true);
           break;
       }
@@ -48,14 +51,12 @@ const LoginScreen = () => {
 
   useEffect(() => {
     findUser(email)
-
   }, [setUserInfo, shouldRedirectToApp]);
 
 
   if (shouldRedirectToRegister) {
-    return (
-      <Redirect push to="/register" />
-    );
+    const navigate = useNavigate()
+    return navigate('/register')
   }
 
   return (
@@ -65,42 +66,44 @@ const LoginScreen = () => {
           <img src={logo} alt="Logo da plataforma" />
           <strong>Acesse sua conta</strong>
           <div className='containerLogin'>
+            <form onSubmit={handleLogin}>
 
-            <div className='inputBox'>
-              <label>E-mail</label>
-              <input
-                type='text'
-                autoFocus
-                required
-                value={email}
-                onChange={(e) => setEmail(e.target.value)}
-              />
-              <p className="errorMsg">{emailError}</p>
-            </div>
-            <div className='inputBox'>
-              <label>Senha</label>
-              <input
-                type="password"
-                required
-                value={password}
-                onChange={(e) => setPassword(e.target.value)}
-              />
-              <p className="errorMsg">{passwordError}</p>
-            </div>
-          </div>
-          <div className="btnContainer">
-            <>
-              {isLoading ? (
-                <ReactLoading className='loading-login-screen-style' type='bars' color='#09629E' height={'20%'} width={'20%'} />
-              ) : (
-                <button onClick={handleLogin} className="BotaoEntrar">Entrar</button>
-              )}
-              <p> Não tem uma conta? <span onClick={() => setShouldRedirectToRegister(true)}>Cadastre-se</span></p>
-            </>
+              <div className='inputBox'>
+                <label>E-mail</label>
+                <input
+                  type='text'
+                  autoFocus
+                  required
+                  value={email}
+                  onChange={(e) => setEmail(e.target.value)}
+                />
+                <p className="errorMsg">{emailError}</p>
+              </div>
+              <div className='inputBox'>
+                <label>Senha</label>
+                <input
+                  type="password"
+                  required
+                  value={password}
+                  onChange={(e) => setPassword(e.target.value)}
+                />
+                <p className="errorMsg">{passwordError}</p>
+              </div>
+              <div className="btnContainer">
+                <>
+                  {isLoading ? (
+                    <ReactLoading className='loading-login-screen-style' type='bars' color='#09629E' height={'20%'} width={'20%'} />
+                  ) : (
+                    <button type='submit' className="BotaoEntrar">Entrar</button>
+                  )}
+                  <p> Não tem uma conta? <span onClick={() => setShouldRedirectToRegister(true)}>Cadastre-se</span></p>
+                </>
+              </div>
+            </form>
           </div>
         </div>
-      </section>
-    </div>
+      </section >
+    </div >
   );
 };
 
