@@ -1,4 +1,4 @@
-import React, { useEffect, useState } from "react";
+import React, { useState } from "react";
 
 import FeedbackCard from "../CardFeedback";
 import { MarginSpace } from "./styles";
@@ -6,13 +6,13 @@ import { useAuth } from "../../firebase";
 import userQueryParams from "../../Views/Perfil/userQueryParams";
 
 export function FeedBacks() {
-   const [avaliacoes,setAvaliacoes] = useState([])
+  const [avaliacoes, setAvaliacoes] = useState([])
   let query = userQueryParams();
 
   const email = query.get("email");
 
-  const { getUserFeedback} = useAuth();
-  
+  const { getUserFeedback } = useAuth();
+
   const getFeedback = () => {
     getUserFeedback(email).then((response) => {
       response.get().then((querySnapshot) => {
@@ -21,38 +21,40 @@ export function FeedBacks() {
           docs = [...docs, { id: doc.id, ...doc.data() }];
         });
 
-        setAvaliacoes(av=>[...docs])
+        setAvaliacoes(av => [...docs])
 
       });
     });
   };
-  useState( ()=>{
+  useState(() => {
 
-      getFeedback()
-  },[])
+    getFeedback()
+  }, [])
   console.log(avaliacoes)
- 
-    return (
-<>
 
-{
-  avaliacoes.length==0?(
+  return (
     <>
-    <h2>Ainda nao foram feitas avaliacoes para esse seguranca</h2>
+
+      {
+        avaliacoes.length == 0 ? (
+          <>
+            <h2>Ainda nao foram feitas avaliacoes para esse seguranca</h2>
+          </>
+        ) :
+
+          avaliacoes.map(dado => {
+            return (
+
+              <MarginSpace>
+                <FeedbackCard author={dado.author}
+                  feedback={dado.feedback}
+                  points={dado.points} >
+
+                </FeedbackCard>
+              </MarginSpace>
+            )
+          })
+      }
     </>
-  ):
-    
-avaliacoes.map(dado =>{
-    return (
-        
-        <MarginSpace>
-<FeedbackCard   author={dado.author}
-    feedback={dado.feedback}
-    points={dado.points} ></FeedbackCard>
-    </MarginSpace>
-    )    
-})
-}
-</>
-    )
+  )
 }
